chore: changed balance column type to double precision

+++ src/main/java/org/wallet/CrudOperations/AccountCrudOperations.java
@@ -55,7 +55,7 @@ public class AccountCrudOperations implements CrudOperations<Account> {
         try {
             PreparedStatement statement = connection.prepareStatement(sql);
 
-            statement.setFloat(1, toSave.getBalance());
+            statement.setDouble(1, toSave.getBalance());
             statement.setString(2, toSave.getCurrencyId());
 
             statement.execute();
@@ -103,7 +103,7 @@ public class AccountCrudOperations implements CrudOperations<Account> {
     private Account mapResultSet(ResultSet resultSet) throws SQLException {
         Account account = new Account();
         account.setAccountId(resultSet.getString("account_id"));
-        account.setBalance(resultSet.getFloat("balance"));
+        account.setBalance(resultSet.getDouble("balance"));
         account.setCurrencyId(resultSet.getString("currency_id"));
 
         return account;
